#902: move sim_review from deprecated (#914)
* #902: move sim_review from deprecated

* #902: update docs
