+++ cdds/bin/cdds_sim_review
@@ -3,7 +3,7 @@
 # Please see LICENSE.md for license details.
 import sys
 
-from cdds.deprecated.transfer.command_line import main_sim_review
+from cdds.sim_review.command_line import main_sim_review
 
 
 if __name__ == '__main__':
