+++ cdds/cdds/deprecated/transfer/command_line.py
@@ -11,14 +11,11 @@ from argparse import Namespace
 from typing import List
 
 from cdds.common import configure_logger, common_command_line_args, check_directory
-from cdds.common.cdds_files.cdds_directories import update_log_dir
 from cdds.common.request.request import read_request
-from cdds.common.plugins.plugin_loader import load_plugin
 
 from cdds import __version__
 from cdds.deprecated.transfer.list_queue import print_queue
 from cdds.deprecated.transfer.resend_failed_msgs import resend_failed_msgs
-from cdds.deprecated.transfer.sim_review import do_sim_review
 from cdds.deprecated.transfer.state import known_states
 from cdds.deprecated.transfer.state_change import run_move_in_mass
 from cdds.deprecated.transfer.constants import KNOWN_RABBITMQ_QUEUES
@@ -107,53 +104,6 @@ def parse_arguments_move_in_mass(arguments: List[str]) -> Namespace:
     if args.output_dir is not None:
         args.output_dir = check_directory(args.output_dir)
     return args
-
-
-def main_sim_review() -> int:
-    """Review the simulation process.
-
-    Returns
-    -------
-    int
-        Exit code
-    """
-    args = parse_sim_review_args()
-
-    request = read_request(args.request)
-    load_plugin(request.metadata.mip_era, request.common.external_plugin, request.common.external_plugin_location)
-
-    # Create the configured logger.
-    configure_logger('sim_review', request.common.log_level, False)
-
-    # Retrieve the logger.
-    logger = logging.getLogger(__name__)
-
-    # Log version.
-    logger.info('Using CDDS Transfer version {}'.format(__version__))
-
-    try:
-        do_sim_review(request, args.request)
-        exit_code = 0
-    except BaseException as exc:
-        exit_code = 1
-        logger.critical(exc, exc_info=PRINT_STACK_TRACE)
-    return exit_code
-
-
-def parse_sim_review_args() -> Namespace:
-    """Parses the command line arguments of the simulation review command
-
-    Returns
-    -------
-    Namespace
-        Command line argument namespace.
-    """
-    log_name = 'sim_review'
-    parser = argparse.ArgumentParser()
-    parser.add_argument('request', help='The location of the request configuration', type=str)
-    common_command_line_args(parser, log_name, logging.INFO, __version__)
-
-    return parser.parse_args()
 
 
 def main_list_queue() -> int:
