+++ cdds/cdds/sim_review/__init__.py
@@ -1,0 +1,2 @@
+# (C) British Crown Copyright 2026, Met Office.
+# Please see LICENSE.md for license details.
